+++ src/compliance_agent/scanner/detectors/frameworks/__init__.py
@@ -3,9 +3,13 @@
 from compliance_agent.scanner.detectors.frameworks.autogen import AutoGenDetector
 from compliance_agent.scanner.detectors.frameworks.base import FrameworkDetector, FrameworkRule
 from compliance_agent.scanner.detectors.frameworks.crewai import CrewAIDetector
+from compliance_agent.scanner.detectors.frameworks.dspy import DSPyDetector
+from compliance_agent.scanner.detectors.frameworks.haystack import HaystackDetector
+from compliance_agent.scanner.detectors.frameworks.instructor import InstructorDetector
 from compliance_agent.scanner.detectors.frameworks.langchain import LangChainDetector
 from compliance_agent.scanner.detectors.frameworks.langgraph import LangGraphDetector
 from compliance_agent.scanner.detectors.frameworks.llamaindex import LlamaIndexDetector
+from compliance_agent.scanner.detectors.frameworks.semantic_kernel import SemanticKernelDetector
 from compliance_agent.scanner.detectors.frameworks.vercel_ai import VercelAIDetector
 
 ALL_FRAMEWORK_DETECTORS: list[type[FrameworkDetector]] = [
@@ -15,16 +19,24 @@ ALL_FRAMEWORK_DETECTORS: list[type[FrameworkDetector]] = [
     LangGraphDetector,
     LlamaIndexDetector,
     VercelAIDetector,
+    SemanticKernelDetector,
+    HaystackDetector,
+    DSPyDetector,
+    InstructorDetector,
 ]
 
 __all__ = [
     "ALL_FRAMEWORK_DETECTORS",
     "AutoGenDetector",
     "CrewAIDetector",
+    "DSPyDetector",
     "FrameworkDetector",
     "FrameworkRule",
+    "HaystackDetector",
+    "InstructorDetector",
     "LangChainDetector",
     "LangGraphDetector",
     "LlamaIndexDetector",
+    "SemanticKernelDetector",
     "VercelAIDetector",
 ]